General inference algorithm

Consider a set of observations sampled from a mixture distribution of a null and an alternative hypothesis. For instance in gene regulation, every observation can correspond to expression levels of a pair of genes wich are sampled from a bivariate normal distribution with zero (null hypothesis) or non-zero (alternative hypothesis) correlation coefficient. In Findr, we predict the probability that any sample follows the alternative hypothesis with the following algorithm (based on and modified from [@Chen:2007]):

For robustness against outliers, we convert every continuous variable into standard normally distributed $N(0,1)$ values using a rank-based inverse normal transformation across all samples. We name this step as supernormalization.

We propose a null and an alternative hypothesis for every likelihood ratio test (LRT) of interest where, by definition, the null hypothesis space is a subset of the alternative hypothesis. Model parameters are replaced with their maximum likelihood estimators (MLEs) to obtain the log likelihood ratio (LLR) between the alternative and null hypotheses.

We derive the analytical expression for the probablity density function (PDF) of the LLR when samples follow the null hypothesis.

We convert LLRs into posterior probabilities of the hypothesis of interest with the empirical estimation of local FDR.

Findr.findr — Function

findr(X::Matrix{T}) where T<:AbstractFloat

Compute posterior probabilities for nonzero pairwise correlations between columns of input matrix X. The probabilities are directed (asymmetric) in the sense that they are estimated from a column-specific background distribution.

findr(dX::T) where T<:AbstractDataFrame

Wrapper for findr(Matrix(dX)) when the input dX is in the form of a DataFrame. The output is then also wrapped in a DataFrame with the same column names.

findr(X::Matrix{T},G::Array{S}) where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero differential expression of colunns of input matrix X across groups defined by one or more categorical variables (columns of G).

Return a matrix of size ncols(X) x ncols(G)

findr(X::Matrix{T},G::Matrix{S},pairGX::Matrix{S}) where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero causal relations between columns of input matrix X. The probabilities are estimated for a subset of columns of X that have a (discrete) instrumental variable in input matrix G. The matching between columns of X and columns of G is given by pairGX, a two-column array where the first column corresponds to a column index in G and the second to a column index in X.

Posterior probabilities are computed for the following tests

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

which can be combined into the mediation test ($P_2 P_3$), the non-independence test ($P_2 P_5$), or Findr's legacy combination ($\frac{1}{2}(P_2 P_5 + P_4)$). Alternatively, individual probability matrices for all tests are returned.

All return matrices have size ncols(X) x ncols(G) .

findr(X1::Matrix{T},X2::Array{T},G::Array{S})  where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero causal relations from columns of input matrix X2 to columns of input matrix X1. The columns of input matrix G are (discrete) instrumental variables for the corresponding columns in X2.

Posterior probabilities are computed for the following tests

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

which can be combined into the mediation test ($P_2 P_3$), the non-independence test ($P_2 P_5$), or Findr's legacy combination ($\frac{1}{2}(P_2 P_5 + P_4)$). Alternatively, individual probability matrices for all tests are returned.

All return matrices have size ncols(X1) x ncols(X2) .