List of package functions

Documentation for Findr.

Distributions.ccdf

Distributions.cdf

Distributions.fit_mle

Distributions.logccdf

Distributions.logpdf

Distributions.pdf

Findr.findr

Findr.findr

Findr.findr

Findr.findr

Findr.findr

Findr.findr

Findr.fit_kde

Findr.fit_mixdist_EM

Findr.fit_mixdist_KDE

Findr.fit_mixdist_mom

Findr.fit_mom

Findr.fit_weighted

Findr.groupmeans

Findr.groupmeans

Findr.llrstats_col

Findr.llrstats_col

Findr.nulldist

Findr.nulllog10pval

Findr.nullpdf

Findr.nullpval

Findr.pi0est

Findr.pprob_col

Findr.pprob_col

Findr.pprob_col

Findr.realLLR_col

Findr.realLLR_col

Findr.realLLR_col

Findr.supernormalize

StatsAPI.fit

StatsAPI.params

Distributions.ccdf — Method

ccdf(d,x)

Evaluate the complementary cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.cdf — Method

cdf(d,x)

Evaluate the cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.fit_mle — Method

fit_mle(LBeta, x)

Fit an LBeta distribution to data x using maximum-likelihood estimation by exploiting its relationship to the Beta distribution.

See https://github.com/JuliaStats/Distributions.jl/blob/master/src/univariate/continuous/beta.jl

Distributions.logccdf — Method

logccdf(d,x)

Evaluate the log complementary cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.logpdf — Method

logpdf(d,x)

Evaluate the probability density function of an LBeta distribution.

Distributions.pdf — Method

pdf(d,x)

Evaluate the probability density function of an LBeta distribution with support on $x\geq 0$.

Findr.findr — Method

findr(dX::T) where T<:AbstractDataFrame

Wrapper for findr(Matrix(dX)) when the input dX is in the form of a DataFrame. The output is then also wrapped in a DataFrame with the same column names.

Findr.findr — Method

findr(X::Matrix{T}) where T<:AbstractFloat

Compute posterior probabilities for nonzero pairwise correlations between columns of input matrix X. The probabilities are directed (asymmetric) in the sense that they are estimated from a column-specific background distribution.

Findr.findr — Method

findr(X1::Matrix{T},X2::Array{T},G::Array{S})  where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero causal relations from columns of input matrix X2 to columns of input matrix X1. The columns of input matrix G are (discrete) instrumental variables for the corresponding columns in X2.

Posterior probabilities are computed for the following tests

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

which can be combined into the mediation test ($P_2 P_3$), the non-independence test ($P_2 P_5$), or Findr's legacy combination ($\frac{1}{2}(P_2 P_5 + P_4)$). Alternatively, individual probability matrices for all tests are returned.

All return matrices have size ncols(X1) x ncols(X2) .

Findr.findr — Method

findr(X::Matrix{T},G::Matrix{S},pairGX::Matrix{S}) where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero causal relations between columns of input matrix X. The probabilities are estimated for a subset of columns of X that have a (discrete) instrumental variable in input matrix G. The matching between columns of X and columns of G is given by pairGX, a two-column array where the first column corresponds to a column index in G and the second to a column index in X.

Posterior probabilities are computed for the following tests

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

which can be combined into the mediation test ($P_2 P_3$), the non-independence test ($P_2 P_5$), or Findr's legacy combination ($\frac{1}{2}(P_2 P_5 + P_4)$). Alternatively, individual probability matrices for all tests are returned.

All return matrices have size ncols(X) x ncols(G) .

Findr.findr — Method

findr(X::Matrix{T},G::Array{S}) where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero differential expression of colunns of input matrix X across groups defined by one or more categorical variables (columns of G).

Return a matrix of size ncols(X) x ncols(G)

Findr.fit_kde — Method

fit_kde(llr)

Fit a distribution function to a vector of log-likelihood ratio values llr using kernel density estimation. To avoid boundary effects in the KDE, the log-likelihoods (which take values in $[0,\infty)$) are first transformed to a vector of z,

$z = \log \left( e^{2 LLR} - 1 \right)$

which takes values in $(-\infty,\infty)$. KDE is applied to z, and the probability density function (pdf) for llr is obtained from the pdf for z by the usual transformation rule for functions of random variables.

Findr.fit_mixdist_EM — Function

fit_mixdist_EM(llr,ns,ng=1,test=:corr; maxiter=1000, tol=1e-3)

Fit a two-component mixture distribution of two LBeta distributions to a vector of log-likelihood ratios llr using an EM algorithm. The first component is the true null distribution for a given Findr test with sample size ns and number of genotype groups ng. The second component is the alternative distribution, assumed to follow an LBeta distribution. The prior probability pi0 of an observation belonging to the null component is fixed and determined by the pi0est function. Hence only the parameters of the alternative component need to be estimated.

The EM algorithm outputs posterior probabilities of the alternative hypothesis being true, in the form of the estimated recognition distribution. The optional parameters maxiter (default value 1000) and tol (default value 1e-3) control the convergence of the EM algorithm.

The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.fit_mixdist_KDE — Function

fit_mixdist_KDE(llr,ns,[ng,test])

Return posterior probabilities for a vector of log-likelihood ratio values llr for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.fit_mixdist_mom — Function

fit_mixdist_mom(llr,ns,ng=1,test=:corr)

Fit a two-component mixture distribution of two LBeta distributions to a vector of log-likelihood ratios llr using a method-of-moments algorithm. The first component is the true null distribution for a given Findr test with sample size ns and number of genotype groups ng. The second component is the alternative distribution, assumed to follow an LBeta distribution. The prior probability pi0 of an observation belonging to the null component is fixed and determined by the pi0est function. Hence only the parameters of the alternative component need to be estimated.

The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.fit_mom — Method

fit_mom(LBeta, m1, m2)

Fit an LBeta distribution to given first and second moments m1 and m2 of the corresponding Beta distribution.

Findr.fit_weighted — Method

fit_weighted(LBeta, x, w)

Fit an LBeta distribution to data x where each observation in x has a weight w between 0 and 1, using the method of moments. The function exploits the relationship to the Beta distribution and is a simple adaptation of the fit function in

https://github.com/JuliaStats/Distributions.jl/blob/master/src/univariate/continuous/beta.jl

Findr.groupmeans — Method

groupmeans(Y,Ycol,E)

Compute the size and mean of each column of matrix Y and of the column vector Ycol for each of the groups (unique values) in categorical vector E.

Findr.groupmeans — Method

groupmeans(Y,E)

Compute the size and mean of each column of matrix Y for each of the groups (unique values) in categorical vector E.

Findr.llrstats_col — Method

llrstats_col(Y,Ycol,E)

Compute the sufficient statistics to compute the log-likelihood ratios for Findr tests 2-5 for a given column vector Ycol (gene) of gene expression values with categorical instrument E against all columns of Y.

Ycol and Y are assumed to have undergone supernormalization with each col having mean zero and variance one. The LLRs are scaled by the number of rows (samples).

The sufficient statistics are:

the covariance ρ between the given Ycol and all columns of Y

the weighted average variances σ1 of each column of matrix Y over the groups (unique values) in E

the weighted average covariance σ2 between Ycol and all columns of Y over the groups (unique values) of E

Findr.llrstats_col — Method

llrstats_col(Y,E)

Compute the sufficient statistics to compute the log-likelihood ratios for Findr tests 2 for a given categorical vector E against all columns of Y.

Y is assumed to have undergone supernormalization with each col having mean zero and variance one. The LLRs are scaled by the number of rows (samples).

The sufficient statistics are the weighted average variances σ1 of each column of matrix Y over the groups (unique values) in E.

Findr.nulldist — Function

nulldist(ns,[ng,test])

Return an LBeta distribution that is the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With only one input argument, the null distribution for the correlation test with ns samples is returned. With two input arguments, or with three arguments and test equal to "corr", the null distribution for the correlation test with ns samples is returned and the second argument is ignored

Findr.nulllog10pval — Function

nulllog10pval(llr,ns,[ng,test])

Return negative log10 p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to "corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.nullpdf — Function

nullpdf(llr,ns,[ng,test])

Return probability distribution function evaluations for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.nullpval — Function

nullpval(llr,ns,[ng,test])

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.pi0est — Method

pi0est(pval)

Estimate the proportion π0 of truly null features in a vector pval of p-values using Storey's method

See also http://varianceexplained.org/files/pi0boot.pdf

Findr.pprob_col — Method

pprob_col(Y,E)

Compute the posterior probabilities for differential expression of columns of matrix Y in the groups defined by categorical vector E using Findr test 2 (Linkage test)

Y is assumed to have undergone supernormalization with each column having mean zero and variance one.

Findr.pprob_col — Method

pprob_col(Y,Ycol,E)

Compute the posterior probabilities for the Findr causal tests for a given column vector Ycol with categorical instrument E against all columns of matrix Y:

- Test 2 (**Linkage test**) 
- Test 3 (**Mediation test**)
- Test 4 (**Relevance test**)
- Test 5 (**Pleiotropy test**)

Y is assumed to have undergone supernormalization with each column having mean zero and variance one.

For test 2, 4, and 5 the posterior probabilities are the probabilities of the alternative hypothesis being true. For test 3 they are the probabilities of the null hypothesis being true.

Findr.pprob_col — Method

pprob_col(Y,Ycol)

Compute the posterior probabilities for Findr test 0 (correlation test) for a given column vector Ycol against all columns of matrix Y.

Y and Ycol are assumed to have undergone supernormalization with each column having mean zero and variance one. The LLRs are scaled by the number of rows (samples).

Findr.realLLR_col — Method

realLLR_col(Y,Ycol,E)

Compute the log-likelihood ratios for Findr test 2 (Linkage test) for a given categorical vector E against all columns of matrix Y.

Y is assumed to have undergone supernormalization with each column having mean zero and variance one. The LLRs are scaled by the number of rows (samples).

Findr.realLLR_col — Method

realLLR_col(Y,Ycol,E)

Compute the log-likelihood ratios for the Findr causal tests for a given column vector Ycol with categorical instrument E against all columns of matrix Y :

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

Y and Ycol are assumed to have undergone supernormalization with each column having mean zero and variance one. The LLRs are scaled by the number of rows (samples).

Findr.realLLR_col — Method

realLLR_col(Y,Ycol)

Compute the log-likelihood ratios for Findr test 0 (correlation test) for a given column vector Ycol against all columns of matrix Y.

Y and Ycol are assumed to have undergone supernormalization with each column having mean zero and variance one. The LLRs are scaled by the number of rows (samples).

Findr.supernormalize — Function

supernormalize(X[, c])

Convert each column of matrix or dataframe X of reals into standard normally distributed values using a rank-based inverse normal transformation. Then scale each column to have variance one.

Note that after the inverse normal transformation, each column has mean zero and identical variance (if we use ordinal ranking). Hence rescaling can be done once on the whole matrix.

The formula and default value for the paramater c come from this paper: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2921808/

StatsAPI.fit — Method

fit(LBeta, x)

Fit an LBeta distribution to data x using the method of moments by exploiting its relationship to the Beta distribution.

See https://github.com/JuliaStats/Distributions.jl/blob/master/src/univariate/continuous/beta.jl

StatsAPI.params — Method

params(d)

Get the parameters of an LBeta distribution.