Utilities

Utility functions not included in any other chapters.

Findr.getpairs — Function

getpairs(dX::T, dG::T, dE::T; colG=1, colX=2)

Get pairs of indices of matching columns from dataframes dX and dG, with column names that should be matched listed in dataframe dE. The optional parameters colG (default value 1) and colX (default value 2) indicate which columns of dE need to be used for matching, either as a column number (integer) or column name (string).

Findr.combineprobs — Function

combineprobs(P; combination="none")

Combine posterior probabilities P for multiple likelihood likelihood ratio tests in a single probability (local precision) value.

Findr.globalfdr — Function

globalfdr(P::Array{T},FDR) where T<:AbstractFloat

For an array (matrix or vector) P of posterior probabilities (local precision values), compute their corresponding q-values Q, and return the indices of P with q-value less than a desired global false discovery rate FDR.

See also qvalue