Findr

Documentation for Findr.

Distributions.ccdf

Distributions.cdf

Distributions.fit_mle

Distributions.logccdf

Distributions.pdf

Findr.fit_kde

Findr.fit_mixdist_EM

Findr.fit_mixdist_KDE

Findr.fit_weighted

Findr.groupmeans

Findr.llrstats_row

Findr.nulldist

Findr.nulllog10pval

Findr.nullpdf

Findr.nullpval

Findr.pi0est

Findr.pprob_corr_row

Findr.realLLRcausal_row

Findr.realLLRcorr_row

Findr.supernormalize

StatsAPI.fit

StatsAPI.params

Distributions.ccdf — Method

ccdf(d,x)

Evaluate the complementary cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.cdf — Method

cdf(d,x)

Evaluate the cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.fit_mle — Method

fit_mle(LBeta, x)

Fit an LBeta distribution to data x using maximum-likelihood estimation by exploiting its relationship to the Beta distribution.

See https://github.com/JuliaStats/Distributions.jl/blob/master/src/univariate/continuous/beta.jl

Distributions.logccdf — Method

logccdf(d,x)

Evaluate the complementary cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.pdf — Method

pdf(d,x)

Evaluate the probability density function of an LBeta distribution with support on $x\geq 0$.

Findr.fit_kde — Method

fit_kde(llr)

Fit a distribution function to a vector of log-likelihood ratio values llr using kernel density estimation. To avoid boundary effects in the KDE, the log-likelihoods (which take values in $[0,\infty)$) are first transformed to a vector of z,

$z = \log \left( e^{2 LLR} - 1 \right)$

which takes values in $(-\infty,\infty)$. KDE is applied to z, and the probability density function (pdf) for llr is obtained from the pdf for z by the usual transformation rule for functions of random variables.

Findr.fit_mixdist_EM — Function

fit_mixdist_EM(llr,pi0,ns,ng=1,test=:corr; maxiter=1000, tol=1e-14)

Fit a two-component mixture distribution of two LBeta distribution to a vector of log-likelihood ratios llr using an EM algorithm. The first component is the true null distribution for a given Findr test with sample size ns and number of genotype groups ng. The second component is the alternative distribution, assumed to follow an LBeta distribution. The prior probability pi0 of an observation belonging to the null component is fixed and determined by the pi0est function. Hence only the parameters of the alternative component need to be estimated.

The EM algorithm outputs posterior probabilities of the alternative hypothesis being true, in the form of the estimated recognition distribution. The optional parameters maxiter (default value 1000) and tol (default value 1e-14) control the convergence of the EM algorithm.

The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.fit_mixdist_KDE — Function

fit_mixdist_KDE(llr,ns,[ng,test])

Return posterior probabilities for a vector of log-likelihood ratio values llr for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.fit_weighted — Method

fit_weighted(LBeta, x, w)

Fit an LBeta distribution to data x where each observation in x has a weight w between 0 and 1, using the method of moments. The function exploits the relationship to the Beta distribution and is a simple adaptation of the fit function in

https://github.com/JuliaStats/Distributions.jl/blob/master/src/univariate/continuous/beta.jl

Findr.groupmeans — Method

groupmeans(Y,E)

Compute the size and mean of each row of matrix Y for each of the groups (unique values) in categorical vector E.

Findr.llrstats_row — Method

llrstats_row(Y,E,row)

Compute the sufficient statistics to compute the log-likelihood ratios for Findr tests 2-5 for a given row (gene) of gene expression matrix Y with categorical instrument E against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

The sufficient statistics are:

the covariance ρ between the given row of matrix Y and all other rows of Y

the weighted average variances σ1 of each row of matrix Y over the groups (unique values) in E

the weighted average covariance σ2 between the given row of Y and all other rows of Y over the groups (unique values) of E

Findr.nulldist — Function

nulldist(ns,[ng,test])

Return an LBeta distribution that is the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With only one input argument, the null distribution for the correlation test with ns samples is returned. With two input arguments, or with three arguments and test equal to "corr", the null distribution for the correlation test with ns samples is returned and the second argument is ignored

Findr.nulllog10pval — Function

nulllog10pval(llr,ns,[ng,test])

Return negative log10 p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to "corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.nullpdf — Function

nullpdf(llr,ns,[ng,test])

Return probability distribution function evaluations for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.nullpval — Function

nullpval(llr,ns,[ng,test])

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for a given Findr test with sample size ns and number of genotype groups ng. The input variable test can take the values:

':corr' - correlation test (test 0)

':link' - linkage test (test 1/2)

':med' - mediation test (test 3)

':relev' - relevance test (test 4)

':pleio' - pleiotropy test (test 5)

With two input arguments, the correlation test with ns samples is used. With three input arguments, or with four arguments and test equal to ":corr", the correlation test with ns samples is used and the third argument is ignored.

Findr.pi0est — Method

pi0est(pval)

Estimate the proportion π0 of truly null features in a vector pval of p-values using Storey's method

See also http://varianceexplained.org/files/pi0boot.pdf

Findr.pprob_corr_row — Method

pprob_corr_row(Y,row)

Compute the posterior probabilities for Findr test 0 (correlation test) for a given row (gene) of gene expression matrix Y against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.realLLRcausal_row — Method

realLLRcausal_row(Y,E,row)

Compute for a given row (gene) of gene expression matrix Y with categorical instrument E against all other rows of Y the log-likelihood ratios for Findr causal tests:

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.realLLRcorr_row — Method

realLLRcorr_row(Y,row)

Compute the log-likelihood ratios for Findr test 0 (correlation test) for a given row (gene) of gene expression matrix Y against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.supernormalize — Function

supernormalize(X[, c])

Convert each row of matrix X of reals into standard normally distributed values using a rank-based inverse normal transformation. Then scale each row to have variance one.

Note that after the inverse normal transformation, each row has mean zero and identical variance (if we use ordinal ranking). Hence rescaling can be done once on the whole matrix.

The formula and default value for the paramater c come from this paper: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2921808/

StatsAPI.fit — Method

fit(LBeta, x)

Fit an LBeta distribution to data x using the method of moments by exploiting its relationship to the Beta distribution.

See https://github.com/JuliaStats/Distributions.jl/blob/master/src/univariate/continuous/beta.jl

StatsAPI.params — Method

params(d)

Get the parameters of an LBeta distribution.