Findr

Documentation for Findr.

Findr.LBeta

Distributions.ccdf

Distributions.cdf

Distributions.logccdf

Distributions.pdf

Findr.groupmeans

Findr.llrstats_row

Findr.randomLLRcorr_log10pval

Findr.randomLLRcorr_pval

Findr.randomLLRlink_log10pval

Findr.randomLLRlink_pval

Findr.randomLLRmed_log10pval

Findr.randomLLRmed_pval

Findr.randomLLRpleio_log10pval

Findr.randomLLRpleio_pval

Findr.randomLLRrelev_log10pval

Findr.randomLLRrelev_pval

Findr.realLLRcausal_row

Findr.realLLRcorr_row

Findr.supernormalize

StatsAPI.params

Findr.LBeta — Type

LBeta(α,β)

The LBeta distribution with parameters $\alpha$ and $\beta$ is defined as the distribution of a random variable

$X=-\frac{1}{2}(\ln(1-Y))$

where $Y\sim\operatorname{Beta}(\alpha, \beta)$

Distributions.ccdf — Method

ccdf(d,x)

Evaluate the complementary cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.cdf — Method

cdf(d,x)

Evaluate the cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.logccdf — Method

logccdf(d,x)

Evaluate the complementary cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.pdf — Method

pdf(d,x)

Evaluate the probability density function of an LBeta distribution with support on $x\geq 0$.

Findr.groupmeans — Method

groupmeans(Y,E)

Compute the size and mean of each row of matrix Y for each of the groups (unique values) in categorical vector E.

Findr.llrstats_row — Method

llrstats_row(Y,E,row)

Compute the sufficient statistics to compute the log-likelihood ratios for Findr tests 2-5 for a given row (gene) of gene expression matrix Y with categorical instrument E against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

The sufficient statistics are:

the covariance ρ between the given row of matrix Y and all other rows of Y

the weighted average variances σ1 of each row of matrix Y over the groups (unique values) in E

the weighted average covariance σ2 between the given row of Y and all other rows of Y over the groups (unique values) of E

Findr.randomLLRcorr_log10pval — Method

randomLLRcorr_log10pval(llr,ns)

Return -log10(p-values) for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 0 (correlation test) with sample size ns.

Findr.randomLLRcorr_pval — Method

randomLLRcorr_pval(llr,ns)

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 0 (correlation test) with sample size ns.

Findr.randomLLRlink_log10pval — Method

randomLLRlink_log10pval(llr,ns,ng)

Return -log10(p-values) for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 2 (linkage test) with sample size ns.

Findr.randomLLRlink_pval — Method

randomLLRlink_pval(llr,ns)

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 2 (linkage test) with sample size ns and number of genotype groups ng.

Findr.randomLLRmed_log10pval — Method

randomLLRmed_log10pval(llr,ns,ng)

Return -log10(p-values) for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 3 (mediation test) with sample size ns.

Findr.randomLLRmed_pval — Method

randomLLRmed_pval(llr,ns)

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 3 (mediation test) with sample size ns and number of genotype groups ng.

Findr.randomLLRpleio_log10pval — Method

randomLLRpleio_log10pval(llr,ns,ng)

Return -log10(p-values) for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 4 (relevance test) with sample size ns.

Findr.randomLLRpleio_pval — Method

randomLLRpleio_pval(llr,ns)

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 4 (relevance test) with sample size ns and number of genotype groups ng.

Findr.randomLLRrelev_log10pval — Method

randomLLRrelev_log10pval(llr,ns,ng)

Return -log10(p-values) for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 4 (relevance test) with sample size ns.

Findr.randomLLRrelev_pval — Method

randomLLRrelev_pval(llr,ns)

Return p-values for a vector of log-likelihood ratio values llr under the null distribution of the log-likelihood ratio for Findr test 4 (relevance test) with sample size ns and number of genotype groups ng.

Findr.realLLRcausal_row — Method

realLLRcausal_row(Y,E,row)

Compute for a given row (gene) of gene expression matrix Y with categorical instrument E against all other rows of Y the log-likelihood ratios for Findr causal tests:

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.realLLRcorr_row — Method

realLLRcorr_row(Y,row)

Compute the log-likelihood ratios for Findr test 0 (correlation test) for a given row (gene) of gene expression matrix Y against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.supernormalize — Function

supernormalize(X[, c])

Convert each row of matrix X of reals into standard normally distributed values using a rank-based inverse normal transformation. Then scale each row to have variance one.

Note that after the inverse normal transformation, each row has mean zero and identical variance (if we use ordinal ranking). Hence rescaling can be done once on the whole matrix.

The formula and default value for the paramater c come from this paper: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2921808/

StatsAPI.params — Method

params(d)

Get the parameters of an LBeta distribution.