Findr

Documentation for Findr.

Findr.LBeta

Distributions.cdf

Distributions.pdf

Findr.randomLLR0

Findr.randomLLR1

Findr.randomLLR3

Findr.randomLLR4

Findr.randomLLR5

Findr.supernormalize

StatsAPI.params

Findr.LBeta — Type

LBeta(α,β)

The LBeta distribution with parameters $\alpha$ and $\beta$ is defined as the distribution of a random variable

$X=-\frac{1}{2}(\ln(1-Y))$

where $Y\sim\operatorname{Beta}(\alpha, \beta)$

Distributions.cdf — Method

cdf(d,x)

Evaluate the cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.pdf — Method

pdf(d,x)

Evaluate the probability density function of an LBeta distribution with support on $x\geq 0$.

Findr.randomLLR0 — Method

randomLLR0(ns)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 0 with sample size ns.

Findr.randomLLR1 — Method

randomLLR1(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 1 with sample size ns and number of genotype groups ng.

Findr.randomLLR3 — Method

randomLLR3(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 3 with sample size ns and number of genotype groups ng.

Findr.randomLLR4 — Method

randomLLR4(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 4 with sample size ns and number of genotype groups ng.

Findr.randomLLR5 — Method

randomLLR5(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 5 with sample size ns and number of genotype groups ng.

Findr.supernormalize — Function

supernormalize(X[, c])

Convert each row of matrix X of reals into standard normally distributed values using a rank-based inverse normal transformation. Then scale each row to have variance one.

Note that after the inverse normal transformation, each row has mean zero and identical variance (if we use ordinal ranking). Hence rescaling can be done once on the whole matrix.

The formula and default value for the paramater c come from this paper: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2921808/

StatsAPI.params — Method

params(d)

Get the parameters of an LBeta distribution.