Findr

Documentation for Findr.

Findr.LBeta

Distributions.cdf

Distributions.pdf

Findr.cov_row

Findr.groupcov

Findr.groupmeans

Findr.groupsizes

Findr.groupvar

Findr.llrstats_row

Findr.randomLLRRrelev

Findr.randomLLRcorr

Findr.randomLLRlink

Findr.randomLLRmed

Findr.randomLLRpleio

Findr.realLLRcorr_row

Findr.realLLRcorr_row

Findr.realLLRlink_row

Findr.realLLRlink_row

Findr.realLLRmed_row

Findr.realLLRpleio_row

Findr.realLLRrelev_row

Findr.supernormalize

StatsAPI.params

Findr.LBeta — Type

LBeta(α,β)

The LBeta distribution with parameters $\alpha$ and $\beta$ is defined as the distribution of a random variable

$X=-\frac{1}{2}(\ln(1-Y))$

where $Y\sim\operatorname{Beta}(\alpha, \beta)$

Distributions.cdf — Method

cdf(d,x)

Evaluate the cumulative distribution function of an LBeta distribution using its relation to the Beta distribution.

Distributions.pdf — Method

pdf(d,x)

Evaluate the probability density function of an LBeta distribution with support on $x\geq 0$.

Findr.cov_row — Method

cov_row(Y,row)

Compute the covariance between a given row of matrix Y and all other rows of Y.

Findr.groupcov — Method

groupcov_row(Y,E,row)

Compute the weighted average covariance between a given row of matrix Y and all other rows of Y over the groups defined by the unique values of vector E.

Findr.groupmeans — Method

groupmeans(Y,E)

Compute the mean of each row of matrix Y for each of the groups (unique values) in categorical vector E.

Findr.groupsizes — Method

groupsizes(E)

Compute the group sizes (number of elements) for each group (unique value) in categorical vector E.

Findr.groupvar — Method

groupvar(Y,E)

Compute the weighted average variance of each row of matrix Y over the groups defined by the unique values of vector E.

Findr.llrstats_row — Method

llrstats_row(Y,E,row)

Compute the sufficient statistics to compute the log-likelihood ratios for Findr tests 2-5 for a given row (gene) of gene expression matrix Y with categorical instrument E against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

The sufficient statistics are:

the group sizes (number of elements) for each group (unique value) in E

the mean of each row of matrix Y for each of the groups (unique values) in E

the weighted average variance of each row of matrix Y over the groups (unique values) in E

the weighted average covariance between the given row of Y and all other rows of Y over the groups (unique values) of E

Findr.randomLLRRrelev — Method

randomLLRRrelev(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 4 (relevance test) with sample size ns and number of genotype groups ng.

Findr.randomLLRcorr — Method

randomLLRcorr(ns)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 0 (correlation test) with sample size ns.

Findr.randomLLRlink — Method

randomLLRlink(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 1/2 (linkage test) with sample size ns and number of genotype groups ng.

Findr.randomLLRmed — Method

randomLLRmed(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 3 (mediation test) with sample size ns and number of genotype groups ng.

Findr.randomLLRpleio — Method

randomLLRpleio(ns,ng)

Return an LBeta distributed random variable for the null distribution of the log-likelihood ratio for Findr test 5 (pleiotropy test) with sample size ns and number of genotype groups ng.

Findr.realLLRcorr_row — Method

realLLRcorr_row(Y,row)

Compute the log-likelihood ratios for Findr test 0 (correlation test) for a given row (gene) of gene expression matrix Y against all other rows of Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.realLLRcorr_row — Method

realLLRcorr_row(ρ)

Compute the log-likelihood ratios for Findr test 0 (correlation test) with precomputed correlation values ̢ρ

Findr.realLLRlink_row — Method

realLLRlink_row(Y,E)

Compute the log-likelihood ratios for Findr test 2 (linkage test) for categorical vector E against all rows of matrix Y.

Y is assumed to have undergone supernormalization with each row having mean zero and variance one. The LLRs are scaled by the number of samples.

Findr.realLLRlink_row — Method

realLLRlink_row(σ)

Compute the log-likelihood ratios for Findr test 2 (linkage test) with precomputed weighted average group variances σ.

Findr.realLLRmed_row — Method

realLLRmed_row(ρ,σ1,σ2)

Compute the log-likelihood ratios for Findr test 3 (mediation test) with precomputed correlation values ρ and weighted average group variances σ1 and covariances σ2

Findr.realLLRpleio_row — Method

realLLRpleio_row(ρ,σ1,σ2)

Compute the log-likelihood ratios for Findr test 3 (mediation test) with precomputed correlation values ρ and weighted average group variances σ1 and covariances σ2

Findr.realLLRrelev_row — Method

realLLRrelev_row(ρ,σ1,σ2)

Compute the log-likelihood ratios for Findr test 3 (mediation test) with precomputed correlation values ρ and weighted average group variances σ1 and covariances σ2

Findr.supernormalize — Function

supernormalize(X[, c])

Convert each row of matrix X of reals into standard normally distributed values using a rank-based inverse normal transformation. Then scale each row to have variance one.

Note that after the inverse normal transformation, each row has mean zero and identical variance (if we use ordinal ranking). Hence rescaling can be done once on the whole matrix.

The formula and default value for the paramater c come from this paper: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2921808/

StatsAPI.params — Method

params(d)

Get the parameters of an LBeta distribution.