General inference algorithm

Consider a set of observations sampled from a mixture distribution of a null and an alternative hypothesis. For instance in gene regulation, every observation can correspond to expression levels of a pair of genes wich are sampled from a bivariate normal distribution with zero (null hypothesis) or non-zero (alternative hypothesis) correlation coefficient. In Findr, we predict the probability that any sample follows the alternative hypothesis with the following algorithm (based on and modified from [Chen2007]):

For robustness against outliers, we convert every continuous variable into standard normally distributed $N(0,1)$ values using a rank-based inverse normal transformation across all samples. We name this step as supernormalization.

We propose a null and an alternative hypothesis for every likelihood ratio test (LRT) of interest where, by definition, the null hypothesis space is a subset of the alternative hypothesis. Model parameters are replaced with their maximum likelihood estimators (MLEs) to obtain the log likelihood ratio (LLR) between the alternative and null hypotheses.

We derive the analytical expression for the probablity density function (PDF) of the LLR when samples follow the null hypothesis.

We convert LLRs into posterior probabilities of the hypothesis of interest with the empirical estimation of local FDR.

While primarily developed for causal inference, several common tasks in genome-wide studies can be implemented using this algorithm and are supported in Findr.

Coexpression analysis

Findr.findr — Method

findr(X::Matrix{T}) where T<:AbstractFloat

Compute posterior probabilities for nonzero pairwise correlations between columns of input matrix X. The probabilities are directed (asymmetric) in the sense that they are estimated from a column-specific background distribution.

See also findr(dX::DataFrame)-

Findr.findr — Method

findr(dX::T) where T<:AbstractDataFrame

Wrapper for findr(Matrix(dX)) when the input dX is in the form of a DataFrame. The output is then also wrapped in a DataFrame with the same column names.

Differential expression analysis

Findr.findr — Method

findr(X::Matrix{T},G::Array{S}) where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero differential expression of colunns of input matrix X across groups defined by one or more categorical variables (columns of G).

Return a matrix of size ncols(X) x ncols(G)

See also findr(dX::DataFrame,dG::DataFrame).

Note

G is currently assumed to be an array (vector or matrix) of integers. CategoricalArrays will be supported in the future.

Findr.findr — Method

findr(dX::T, dG::T) where T<:AbstractDataFrame

Wrapper for findr(Matrix(dX), Matrix(dG)) when the inputs dX and dG are in the form of a DataFrame. The output is then also wrapped in a DataFrame with the same column names.

Causal inference

Findr.findr — Method

findr(X::Matrix{T},G::Matrix{S},pairGX::Matrix{S}; combination="none") where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero causal relations between columns of input matrix X. The probabilities are estimated for a subset of columns of X that have a (discrete) instrumental variable in input matrix G. The matching between columns of X and columns of G is given by pairGX, a two-column array where the first column corresponds to a column index in G and the second to a column index in X.

Posterior probabilities are computed for the following tests

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

which can be combined into the mediation test ($P_2 P_3$; combination="mediation"), the instrumental variable or non-independence test ($P_2 P_5$; combination="IV"), or Findr's original combination ($\frac{1}{2}(P_2 P_5 + P_4)$; combination="orig"). By default, individual probability matrices for all tests are returned (combination="none").

All return matrices have size ncols(X) x ncols(G).

Note

G is currently assumed to be an array (vector or matrix) of integers. I intend to use CategoricalArrays in the future.

Findr.findr — Method

findr(dX::T, dG::T, dE::T, colX=2, colG=1, combination="IV") where T<:AbstractDataFrame

Wrapper for findr(X, G, pairGX) when the inputs are in the form of a DataFrame. The output is then also wrapped in a DataFrame with the same column names. When DataFrames are used, only a combined posterior probabilities can be returned (combination="IV" (default), "mediation", or "orig").

The input dataframes are:

dX - DataFrame with expression data, columns are genes

dG - DataFrame with genotype data, columns are variants (SNPs)

dE - DataFrame with eQTL results, must contains columns with gene and SNP IDs that can be mapped to column names in dX and dG, respectively

colG - name or number of variant ID column in dE, default 1

colX - name or number of gene ID column in dE, default 2

Bipartite causal inference

Findr.findr — Method

findr(X1::Matrix{T},X2::Array{T},G::Array{S},pairGX::Matrix{R}; combination="none")  where {T<:AbstractFloat, S<:Integer}

Compute posterior probabilities for nonzero causal relations from columns of input matrix X2 to columns of input matrix X1. The probabilities are estimated for a subset of columns of X2 that have a (discrete) instrumental variable in input matrix G. The matching between columns of X2 and columns of G is given by pairGX, a two-column array where the first column corresponds to a column index in G and the second to a column index in X2.

Posterior probabilities are computed for the following tests

Test 2 (Linkage test)

Test 3 (Mediation test)

Test 4 (Relevance test)

Test 5 (Pleiotropy test)

which can be combined into the mediation test ($P_2 P_3$; combination="mediation"), the instrumental variable or non-independence test ($P_2 P_5$; combination="IV"), or Findr's original combination ($\frac{1}{2}(P_2 P_5 + P_4)$; combination="orig"). By default, individual probability matrices for all tests are returned (combination="none").

All return matrices have size ncols(X1) x ncols(X2).

Note

G is currently assumed to be an array (vector or matrix) of integers. I intend to use CategoricalArrays in the future.

Findr.findr — Method

findr(dX1::T, dX2::T, dG::T, dE::T; colG=1, colX=2, combination="IV") where T<:AbstractDataFrame

Wrapper for findr(Matrix(dX1), Matrix(dX2), Matrix(dG)) when the inputs dX1, dX2, and dG are in the form of a DataFrame. The output is then also wrapped in a DataFrame with the same column names. When DataFrames are used, only a combined posterior probabilities can be returned (combination="IV" (default), "mediation", or "orig").

Chen2007Chen L, Emmert-Streib F, Storey J. Harnessing naturally randomized transcription to infer regulatory relationships among genes. Genome Biol 8, R219 (2007).